requirements.txt
absl-py==0.15.0
annotated-types==0.7.0
anyio==4.5.2
argcomplete==3.6.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==25.3.0
Automat==24.8.1
av==12.3.0
av2==0.2.1
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.1.0
buildtools==1.0.6
cachetools==5.5.2
causal-conv1d==1.1.3.post1
ccimport==0.4.4
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
cleo==2.1.0
click==8.1.8
colorlog==6.9.0
comm==0.2.2
constantly==23.10.4
contourpy==1.1.1
crashtest==0.4.1
cryptography==45.0.4
cumm-cu116==0.4.11
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dependency-groups==1.3.1
descartes==1.1.0
distlib==0.3.9
dnspython==2.6.1
docopt==0.6.2
easydict==1.13
einops==0.8.1
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
fire==0.7.0
flatbuffers==1.12
fonttools==4.57.0
fqdn==1.5.1
fsspec==2024.6.1
furl==2.1.4
gast==0.3.3
glog==0.3.1
google==3.0.0
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.32.0
h11==0.16.0
h5py==2.10.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
hyperlink==21.0.0
idna==3.4
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
incremental==24.7.2
iniconfig==2.1.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
jwt==1.3.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
kornia==0.6.8
kornia_rs==0.1.8
lark==1.2.2
lazy_loader==0.4
llvmlite==0.31.0
mamba-ssm==1.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
motmetrics==1.4.0
mpmath==1.3.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.0
ninja==1.11.1.4
notebook==7.3.3
notebook_shim==0.2.4
nox==2025.5.1
numba==0.48.0
numpy==1.23.3
nuscenes-devkit==1.0.5
oauthlib==3.2.2
opencv-python==4.11.0.86
opt-einsum==3.3.0
orderedmultidict==1.0.1
overrides==7.7.0
packaging==25.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pccm==0.4.16
-e git+https://github.com/zhanggang001/HEDNet.git@8694e7f07e860416bd99dbf46f952b3e6b6c0c6f#egg=pcdet
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pluggy==1.5.0
polars==1.8.2
portalocker==3.0.0
prettytable==2.2.1
prometheus_client==0.21.1
prompt_toolkit==3.0.51
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pycparser==2.22
pycryptodome==3.9.9
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.1.4
pyproj==3.5.0
pyquaternion==0.9.9
pytest==8.3.5
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-gflags==3.1.2
python-json-logger==3.3.0
-e git+https://github.com/facebookresearch/pytorch3d.git@71db7a0ea293f2626f5ea1c70366870a607129d0#egg=pytorch3d
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==27.0.0
rapidfuzz==3.9.7
redo==3.0.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.20.1
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.3
Shapely==1.8.4
SharedArray==3.0.0
simplejson==3.20.1
six==1.15.0
sniffio==1.3.1
soupsieve==2.7
spconv-cu116==2.3.6
SQLAlchemy==2.0.41
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.15
tinycss2==1.2.1
tjob==0.7.5
tokenizers==0.20.3
tomli==2.2.1
torch==2.1.0+cu118
torch-scatter @ file:///mnt/csi-data-aly/user/hongfeizhang/mypaper/SparseFormerV2/torch_scatter-2.1.2%2Bpt21cu118-cp38-cp38-linux_x86_64.whl#sha256=adea2c960381e0623dd0e46a2c61dde585f3cc2b4bed05ea92af1e27a7dbd1cf
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
triton==2.1.0
Twisted==24.11.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.2
universal_pathlib==0.2.6
uri-template==1.3.0
urllib3==1.26.13
virtualenv==20.31.2
volcengine==1.0.191
waymo-open-dataset-tf-2-4-0==1.4.1
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.14
wrapt==1.12.1
xmltodict==0.14.2
zipp==3.20.2
zope.interface==7.2
